locust_tasks/requirements.txt
pyyaml
hvac
locust-plugins
cryptography